backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
stripe==8.2.0
requests==2.31.0
celery==5.3.6
redis==5.0.1
drf-spectacular==0.27.1
gunicorn==21.2.0
